Add targetPath to MockDescription for MockGenerator to be able to pick up original modue path
